rename lib/hook folder to lib/hooks; move useGetRandomPieceOnInterval to lib/hooks
