+++ components/CommentForm/CommentForm.js
@@ -1,4 +1,4 @@
-import { useImmerLocalStorageState } from "../../lib/hook/useImmerLocalStorageState";
+import { useImmerLocalStorageState } from "../../lib/hooks/useImmerLocalStorageState";
 
 export function CommentForm({ title }) {
   const [artPiecesComments, updateArtPiecesComments] =
