+++ components/SpotLight/SpotLight.js
@@ -1,33 +1,10 @@
 import { useState, useEffect } from "react";
 import { ArtPiecePreview } from "..";
 import { getRandomNumber } from "../../lib/utils";
-
-export function useGetRandomPieceOnIntervall(
-  pieces,
-  getRandomNumber,
-  setRandomIndex
-) {
-  useEffect(() => {
-    const intervalID = setInterval(() => {
-      setRandomIndex(getRandomNumber(pieces));
-    }, 3000);
-    return () => {
-      clearInterval(intervalID);
-    };
-  });
-}
+import useGetRandomPieceOnIntervall from "../../lib/hooks/useGetRandomPieceOnIntervall";
 
 export function SpotLight({ pieces, handleToggleFavorite, artPiecesInfo }) {
   const [randomIndex, setRandomIndex] = useState(getRandomNumber(pieces));
-
-  // useEffect(() => {
-  //   const intervalID = setInterval(() => {
-  //     setRandomIndex(getRandomNumber(pieces));
-  //   }, 3000);
-  //   return () => {
-  //     clearInterval(intervalID);
-  //   };
-  // });
 
   useGetRandomPieceOnIntervall(pieces, getRandomNumber, setRandomIndex);
